refactored by putting handlepostclick function in helpers file
